% Auto-generated: do not edit by hand
\name{dfResizableNode}

\alias{dfResizableNode}

\title{ResizableNode component}

\description{
ResizableNode is a custom node component that supports resizing and Dash components.
}

\usage{
dfResizableNode(data=NULL, selected=NULL)
}

\arguments{
\item{data}{Lists containing elements 'label'.
those elements have the following types:
  - label (logical | numeric | character | named list | unnamed list; optional): the content to display in the node. can be a string or dash component.. The node data containing the content to display}

\item{selected}{Logical. Whether the node is currently selected}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

